Fixed bug in ChromeConnection.requestListeners
The ChromeConnection.requestListeners collection was attached to the
prototype and therefore being shared by all instances of the class.
This instantiates  the collection on the instance.

This manifested itself by displaying the request data for every request
in all instances of the inspector instead of only the tab that
originated it